pyproject.toml
[tool.poetry]
name = "cost-manager"
version = "0.1.0"
description = ""
authors = ["MattheusBy <mattaudia6@gmail.com>"]
readme = "README.md"
packages = [{include = "cost_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
drf-registration = "^0.1.3"
postgres = "^4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
